requirements.txt
pydantic~=2.11.7
h11~=0.14.0
anyio~=4.10.0
typer~=0.9.0
python-dotenv~=1.1.1
mcp~=1.12.4
httpx~=0.27.0
httpx-sse~=0.4.0
jsonschema~=4.23.0
websockets~=15.0.1
starlette~=0.46.2
sse-starlette~=3.0.2
pydantic-settings~=2.6.1
jupyter~=1.0.0
rich~=13.7.1
panel~=1.5.2
attrs~=23.1.0
pygments~=2.15.1
zlib~=1.2.13
markdown-it-py~=2.2.0
sniffio~=1.3.0
pytest~=7.4.4
gettext~=0.21.0
click~=8.2.1
httpcore~=1.0.2
idna~=3.7
brotli~=1.0.9
zstandard~=0.23.0
mdurl~=0.1.0
shellingham~=1.5.0
ipython~=8.27.0
certifi~=2025.4.26
uvloop~=0.21.0
uvicorn~=0.34.3
httptools~=0.6.4
watchfiles~=1.1.0
mypy~=1.11.2
annotated-types~=0.6.0
pillow~=10.4.0
docutils~=0.18.1
sphinx~=7.3.7
itsdangerous~=2.2.0
yaml~=0.2.5
pyyaml~=6.0.1
jinja2~=3.1.4
referencing~=0.30.2
webcolors~=24.11.1
jsonpointer~=2.1
uri-template~=1.3.0
isoduration~=20.11.0
fqdn~=1.5.1
rfc3339-validator~=0.1.4
jsonschema-specifications~=2023.7.1
linkify-it-py~=2.0.0
python-multipart~=0.0.20
protobuf~=4.25.3
tomli~=2.0.1
toml~=0.10.2